feat: prefer treesitter over semantic tokens for variables
Disable use of semantic token for variable in favor of treesitter. I
grabbed this from Tokyonight. It allows for different highlight groups
for things such as this Python code object.attr. I set the
variable.member to white, so it is slightly different from the fg used
on the object variable.