Add support for mini, leap, treesitter-context, and two extras (fzf, lazygit) (#53)
* feat: add lazygit theme

* feat: add fzf (and zoxide) theme

* feat: add contrast to status line

Use gray02 instead of gray01 for three reasons. It provides better
contrast between windows. It matches the color used by WinSeparator, so
frames around windows use the same color. It provides better contrast
with the cursor line.

* feat: add contrast to matches in cmp and blink

Use cyan instead of blue for better contrast with matched portions and
non-matched text. As part of this, I also set BlinkCmpLabelMatch to use
cyan as well for the same reasons. I will also use this color in
mini.pick as well.

* feat: use blue for diagnostic hints

Use blue for hints instead of cyan, which is a bright red and could be
mistaken for error messages.

* feat: prefer treesitter over semantic tokens for variables

Disable use of semantic token for variable in favor of treesitter. I
grabbed this from Tokyonight. It allows for different highlight groups
for things such as this Python code object.attr. I set the
variable.member to white, so it is slightly different from the fg used
on the object variable.

* feat: add mini.nvim support

I use almost all of mini.nvim and these are the only changes I needed to
tweak for a more optimal Mellow experience. The only mini plugins I
don't use are mine.pairs, mini.jump2d, and mini.completion.

* feat: add leap support

* feat: add treesitter-context support